# swagger.yaml
openapi: 3.0.3
info:
  title: PDF to PowerPoint Converter API
  description: |
    A FastAPI application for creating PDF files and converting them to PowerPoint presentations using MVC architecture.
    
    ## Features
    - **PDF Creation**: Generate PDF files from text content with custom styling
    - **PDF to PowerPoint Conversion**: Convert PDF files to PowerPoint presentations
    - **File Management**: Upload, download, list, and delete files
    - **RESTful API**: Clean and intuitive API endpoints
    
    ## Authentication
    This API currently does not require authentication for demonstration purposes.
    In production, consider implementing JWT tokens or API keys.
    
    ## Rate Limiting
    No rate limiting is currently implemented. Consider adding rate limiting for production use.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /pdf/create:
    post:
      tags:
        - PDF Operations
      summary: Create PDF
      description: Create a PDF file from provided content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PDFContent'
            examples:
              basic:
                summary: Basic PDF Creation
                value:
                  title: "Sample Document"
                  content: "This is a sample document created via the API."
                  author: "John Doe"
                  subject: "Sample Subject"
                  keywords: ["sample", "document", "pdf"]
              detailed:
                summary: Detailed PDF with Metadata
                value:
                  title: "Technical Report"
                  content: |
                    This is a comprehensive technical report.
                    
                    It includes multiple paragraphs and detailed information.
                    
                    Key Features:
                    • Professional formatting
                    • Custom styling
                    • Metadata support
                  author: "Technical Writer"
                  subject: "Technical Documentation"
                  keywords: ["technical", "report", "documentation", "api"]
      responses:
        '200':
          description: PDF created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PDFResponse'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pdf/list:
    get:
      tags:
        - PDF Operations
      summary: List PDF Files
      description: Get a list of all available PDF files
      responses:
        '200':
          description: PDF files retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  pdfs:
                    type: array
                    items:
                      $ref: '#/components/schemas/PDFFileInfo'
                  count:
                    type: integer
                    example: 2
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pdf/download/{filename}:
    get:
      tags:
        - PDF Operations
      summary: Download PDF
      description: Download a PDF file by filename
      parameters:
        - name: filename
          in: path
          required: true
          description: Name of the PDF file to download
          schema:
            type: string
            example: "Sample_Document_20250815_125355.pdf"
      responses:
        '200':
          description: PDF file downloaded successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: PDF file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pdf/info/{filename}:
    get:
      tags:
        - PDF Operations
      summary: Get PDF Information
      description: Get detailed information about a specific PDF file
      parameters:
        - name: filename
          in: path
          required: true
          description: Name of the PDF file
          schema:
            type: string
            example: "Sample_Document_20250815_125355.pdf"
      responses:
        '200':
          description: PDF information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  info:
                    $ref: '#/components/schemas/PDFFileInfo'
        '404':
          description: PDF file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pdf/delete/{filename}:
    delete:
      tags:
        - PDF Operations
      summary: Delete PDF
      description: Delete a PDF file by filename
      parameters:
        - name: filename
          in: path
          required: true
          description: Name of the PDF file to delete
          schema:
            type: string
            example: "Sample_Document_20250815_125355.pdf"
      responses:
        '200':
          description: PDF file deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "PDF file Sample_Document_20250815_125355.pdf deleted successfully"
        '404':
          description: PDF file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    PDFContent:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: Title of the PDF document
          example: "Sample Document"
          maxLength: 200
        content:
          type: string
          description: Main content of the PDF
          example: "This is the main content of the document..."
        author:
          type: string
          description: Author of the document
          example: "John Doe"
          maxLength: 100
        subject:
          type: string
          description: Subject of the document
          example: "Sample Subject"
          maxLength: 200
        keywords:
          type: array
          items:
            type: string
          description: Keywords for the document
          example: ["sample", "document", "pdf"]
          maxItems: 10
    PDFResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "PDF created successfully: Sample_Document_20250815_125355.pdf"
        pdf_path:
          type: string
          example: "outputs/pdfs/Sample_Document_20250815_125355.pdf"
        created_at:
          type: string
          format: date-time
          example: "2025-08-15T12:53:55.873833"
...
    PDFFileInfo:
      type: object
      properties:
        filename:
          type: string
          example: "Sample_Document_20250815_125355.pdf"
        size_bytes:
          type: integer
          example: 1853
        created_at:
          type: string
          format: date-time
          example: "2025-08-15T12:53:55.873778"
        modified_at:
          type: string
          format: date-time
          example: "2025-08-15T12:53:55.873778"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "PDF file not found"
        message:
          type: string
          example: "The requested PDF file could not be found"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-15T12:54:42.002577"
        path:
          type: string
          example: "/pdf/download/nonexistent.pdf"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (not currently required)
security:
  - ApiKeyAuth: []
...